requirements.txt
amqp==5.1.1
app==0.0.1
appdirs==1.4.4
asgiref==3.3.2
astroid==2.6.2
asttokens==2.0.5
async-timeout==4.0.2
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==21.7b0
celery==5.2.7
certifi==2022.12.7
cfgv==3.3.0
charset-normalizer==2.0.8
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
commonmark==0.9.1
distlib==0.3.2
Django==3.2.14
django-bootstrap-datepicker-plus==5.0.2
django-celery-beat==2.4.0
django-currentuser==0.5.3
django-doctor==2.5.0
django-jazzmin==2.6.0
django-redis==5.2.0
django-timezone-field==5.0
djangorestframework==3.14.0
dparse==0.5.1
filelock==3.0.12
flake8==3.9.2
Flask==2.2.2
flower==1.2.0
humanize==4.4.0
identify==2.2.11
idna==3.3
importlib-metadata==5.2.0
isort==5.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
lazy-object-proxy==1.6.0
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.0
pathspec==0.8.1
Pillow==9.3.0
platformdirs==2.2.0
poetry-semver==0.1.0
pre-commit==2.13.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pycodestyle==2.7.0
pydantic==1.10.2
pyflakes==2.3.1
Pygments==2.9.0
pylint==2.9.3
pyparsing==2.4.7
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7
PyYAML==5.4.1
redis==4.4.0
regex==2021.7.6
requests==2.26.0
rich==10.6.0
six==1.16.0
skjold==0.3.2
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
tomli==1.0.4
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.7
vine==5.0.0
virtualenv==20.7.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.12.1
zipp==3.11.0
